fix CI & renamed files
