+++ website/docs/providers/future_provider.mdx
@@ -1,5 +1,6 @@
 ---
 title: FutureProvider
+version: 1
 ---
 
 import Tabs from "@theme/Tabs";
@@ -7,8 +8,7 @@ import TabItem from "@theme/TabItem";
 import CodeBlock from "@theme/CodeBlock";
 import configProvider from "./future_provider/config_provider";
 import configConsumer from "./future_provider/config_consumer";
-import { trimSnippet,AutoSnippet,
-  ConditionalSnippet, } from "../../src/components/CodeSnippet";
+import { trimSnippet,AutoSnippet} from "../../src/components/CodeSnippet";
 
 `FutureProvider` is the equivalent of [Provider] but for asynchronous code.
 
@@ -38,11 +38,6 @@ syntax, but inside the provider.
 Using Flutter's asset system, this would be:
 
 <AutoSnippet language="dart" {...configProvider}></AutoSnippet>
-
-<ConditionalSnippet codegen={true}>
-  Finally, run the code-generator with{" "}
-  <code>flutter pub run build_runner watch</code>
-</ConditionalSnippet>
 
 
 Then, the UI can listen to configurations like so:
