+++ website/docs/providers/future_provider/config_consumer/codegen.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import '../config_provider/cgen_provider.dart';
+
+import '../config_provider/codegen.dart';
+
 
 /* SNIPPET START */
