+++ website/docs/providers/future_provider/config_consumer/raw.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../config_provider/raw_config_provider.dart';
+import '../config_provider/raw.dart';
 
 /* SNIPPET START */
