+++ website/i18n/es/docusaurus-plugin-content-docs/current/providers/future_provider.mdx
@@ -5,8 +5,8 @@ title: FutureProvider
 import Tabs from "@theme/Tabs";
 import TabItem from "@theme/TabItem";
 import CodeBlock from "@theme/CodeBlock";
-import configProvider from "!!raw-loader!/i18n/ja/docusaurus-plugin-content-docs/current/providers/future_provider/config_provider.dart";
-import configConsumer from "!!raw-loader!/i18n/ja/docusaurus-plugin-content-docs/current/providers/future_provider/config_consumer.dart";
+import configProvider from "!!raw-loader!/i18n/ko/docusaurus-plugin-content-docs/current/providers/future_provider/config_provider.dart";
+import configConsumer from "!!raw-loader!/i18n/ko/docusaurus-plugin-content-docs/current/providers/future_provider/config_consumer.dart";
 import { trimSnippet } from "../../../../../src/components/CodeSnippet";
 
 `FutureProvider` es el equivalente de [Provider] pero para código asíncrono.
