+++ website/i18n/ko/docusaurus-plugin-content-docs/current/providers/stream_provider.mdx
@@ -5,8 +5,8 @@ title: StreamProvider
 import Tabs from "@theme/Tabs";
 import TabItem from "@theme/TabItem";
 import CodeBlock from "@theme/CodeBlock";
-import configProvider from "!!raw-loader!/docs/providers/future_provider/config_provider.dart";
-import configConsumer from "!!raw-loader!/docs/providers/future_provider/config_consumer.dart";
+import configProvider from "/docs/providers/future_provider/config_provider";
+import configConsumer from "/docs/providers/future_provider/config_consumer";
 import { trimSnippet } from "../../../../../src/components/CodeSnippet";
 
 `StreamProvider` is similar to [FutureProvider] but for [Stream]s instead of
