+++ website/src/outdated_translations.js
@@ -1,1 +1,1 @@
-export default [{"countryCode":"bn","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"de","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"es","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"fr","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"it","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"ja","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"ko","id":"getting_started","englishPath":"/docs/getting_started"},]
+export default [{"countryCode":"bn","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"bn","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"de","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"de","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"es","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"es","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"fr","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"fr","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"it","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"it","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"ja","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"ja","id":"future_provider","englishPath":"/docs/providers/future_provider"},{"countryCode":"ko","id":"getting_started","englishPath":"/docs/getting_started"},{"countryCode":"ko","id":"future_provider","englishPath":"/docs/providers/future_provider"},]
